all_metrics/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
spacy
textstat
python-Levenshtein
deep-translator
transliterate
numpy
pandas
pytest
scikit-learn
